requirements.txt
absl-py==2.1.0
aiocache==0.11.1
aiofiles==24.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiopubsub==2.1.5
aioredis==2.0.1
aiosignal==1.3.1
aiosmtplib==1.1.2
alembic==1.3.3
aniso8601==7.0.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.3
appscript==1.2.5
APScheduler==3.6.3
asttokens==2.4.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==23.2.0
ax-platform==0.4.1
backcall==0.2.0
bidict==0.23.1
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
colorednoise==2.2.0
comm==0.1.4
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
diffrax==0.5.1
docopt==0.6.2
docutils==0.19
dotmap==1.3.8
equinox==0.11.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.109.2
filelock==3.13.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gin==0.1.6
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.3
graphql-ws==0.3.0
grpcio==1.65.1
gunicorn==20.0.4
h11==0.14.0
h2==3.2.0
hpack==3.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperframe==5.2.0
idna==2.10
ifaddr==0.2.0
imageio==2.35.1
ipykernel==6.25.2
ipython==8.15.0
ipywidgets==8.1.3
itsdangerous==2.2.0
jaxtyping==0.2.28
jedi==0.19.0
Jinja2==3.1.4
joblib==1.3.2
kiwisolver==1.4.5
linear-operator==0.5.2
lineax==0.0.5
loguru==0.4.1
lxml==5.3.0
Mako==1.3.5
Markdown==3.6
markdown2==2.3.8
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==4.7.6
multipledispatch==1.0.0
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.3
nicegui==1.4.29
numpy==1.25.2
openpyxl==3.1.5
opt-einsum==3.3.0
optax==0.2.2
optimistix==0.0.7
orjson==3.10.6
packaging==20.1
parso==0.8.3
passlib==1.7.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==3.10.0
plotly==5.23.0
promise==2.3
prompt-toolkit==3.0.39
protobuf==4.25.3
pscript==0.7.7
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-postgresql==1.2.1
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.16.1
PyJWT==2.9.0
pyparsing==3.1.2
pyre-extensions==0.0.30
pyro-api==0.1.2
pyro-ppl==1.9.0
pysr==0.19.4
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-editor==1.0.4
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2024.1
PyYAML==5.1.2
pyzmq==25.1.1
redis==5.0.8
replicate==0.28.0
requests==2.32.3
requests-async==0.5.0
rfc3986==1.5.0
rsa==4.0
ruamel.yaml==0.16.7
Rx==1.6.3
sanic==19.6.3
Sanic-Cors==0.9.9.post4
Sanic-GraphQL==1.1.0
sanic-jwt==1.4.0
Sanic-Plugins-Framework==0.9.5
sanic_compress==0.1.1
scikit-learn==1.4.1.post1
scipy==1.13.0
semver==3.0.2
simple-websocket==1.0.0
singledispatch==3.7.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.3.13
SQLAlchemy-Utils==0.36.1
stack-data==0.6.2
starlette==0.36.3
stripe==2.42.0
tenacity==9.0.0
threadpoolctl==3.4.0
toolz==0.12.1
tornado==6.3.3
tqdm==4.66.2
traitlets==5.10.0
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
vbuild==0.8.2
watchfiles==0.22.0
wcwidth==0.2.6
websockets==13.0.1
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wsproto==1.2.0
xlwings==0.32.2
yarl==1.9.4
jaxopt
